Tests should use float32

+++ test/test_convert.py
@@ -12,7 +12,7 @@ import torch
 def test_convert_frequency_to_midi(audio):
     """Test that conversion from Hz to MIDI matches librosa implementation"""
 
-    sample_data = torch.tensor([110.0, 220.0, 440.0, 500.0, 880.0])
+    sample_data = torch.tensor([110.0, 220.0, 440.0, 500.0, 880.0], dtype=torch.float32)
 
     penn_midi = penn.convert.frequency_to_midi(sample_data)
 
@@ -23,7 +23,7 @@ def test_convert_frequency_to_midi(audio):
 def test_convert_midi_to_frequency(audio):
     """Test that conversion from MIDI to Hz matches librosa implementation"""
 
-    sample_data = torch.tensor([45.0, 57.0, 69.0, 71.2131, 81.0])   
+    sample_data = torch.tensor([45.0, 57.0, 69.0, 71.2131, 81.0], dtype=torch.float32)   
 
     penn_frequency = penn.convert.midi_to_frequency(sample_data)
